pyproject.toml
[tool.poetry]
name = "NeuroScan"
version = "0.1.0"
description = "Brain Tumor Detection With CNNs"
authors = ["JavithNaseem-J <javithnaseem.j@gmail.com>"]
packages = [{ include = "NeuroScan", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
pandas = "^2.3.1"
numpy = "^2.0.2"
scikit-learn = "^1.6.1"
matplotlib = "^3.9.4"
seaborn = "^0.13.2"
joblib = "^1.5.1"
tqdm = "^4.67.1"
scipy = "^1.13.1"
imutils = "^0.5.4"
pyyaml = "^6.0.2"
python-box = "^7.3.2"
ensure = "^1.0.4"
python-dotenv = "^1.1.1"
streamlit = "^1.12.0"
optuna = "^4.4.0"
dvc = "^3.61.0"
mlflow = "^3.1.4"
dvc-s3 = "^3.2.2"
boto3 = "^1.39.8"
types-pyyaml = "^6.0.12"
opencv-python = "^4.12.0"
[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ipykernel = "^6.30.0"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"